Move config handling code to extra file
For better code structure.
