+++ jiracli/tests/test_base.py
@@ -17,11 +17,6 @@
 
 
 import unittest
-import mock
-import os
-import shutil
-import stat
-import tempfile
 
 from ddt import ddt, data, unpack
 import jiracli
@@ -40,18 +35,3 @@ class BaseTest(unittest.TestCase):
     @unpack
     def test_issue_status_color(self, status, expected_color):
         assert jiracli.issue_status_color(status) == expected_color
-
-    @mock.patch('jiracli.config_credentials_get',
-                return_value=('joe', 'secret', 'http://jira.example.com'))
-    def test_config_get_file_not_available(self, mock_creds_get):
-        """get a non-available config - file should be created"""
-        tmpdir = tempfile.mkdtemp(prefix='jiracli-tmp_')
-        conf = os.path.join(tmpdir, 'jiracli.conf')
-        try:
-            jiracli.config_get(conf)
-            # file was created
-            assert os.path.exists(conf) is True
-            # file has correct permissions
-            assert oct(stat.S_IMODE(os.lstat(conf).st_mode)) == oct(0o600)
-        finally:
-            shutil.rmtree(tmpdir)
